+++ lib/main.dart
@@ -10,7 +10,7 @@ import 'package:novynaplo/screens/avarages_tab.dart';
 import 'package:novynaplo/screens/events_tab.dart';
 import 'package:novynaplo/screens/marks_tab.dart';
 import 'package:novynaplo/screens/reports_tab.dart';
-import 'package:novynaplo/screens/settings_tab.dart';
+import 'package:novynaplo/screens/settings/settings_tab.dart';
 import 'package:novynaplo/screens/login_page.dart';
 import 'package:novynaplo/screens/notices_tab.dart';
 import 'package:novynaplo/screens/statistics_tab.dart';
